Hu#23 denunciar producto

+++ server/controllers/apiTradeExchange.js
@@ -9,7 +9,6 @@ exports.readAllTradeExchange = async (req, res) => {
   
       res.status(200).json(tradeExchange);
       
-      console.log(tradeExchange);
     } catch (error) {
       res.status(400).json(error.message);
       console.log(error.message);
@@ -32,36 +31,40 @@ exports.readAllTradeExchange = async (req, res) => {
   exports.createTradeExchange = async (req, res) => {
     const tradeExchange = new TradeExchange();
 
-    tradeExchange.userOfering = req.body.userOfering;
+    tradeExchange.userOfering = req.user.id;
     tradeExchange.userTaking = req.body.userTaking;
     tradeExchange.productOfering = req.body.productOfering;
     tradeExchange.productTaking = req.body.productTaking;
-   tradeExchange.publishingDate = req.body.publishingDate;
+    tradeExchange.publishingDate = req.body.publishingDate;
+    tradeExchange.points = req.body.points; 
      
     try {
-        const userOfering = await User.findById({_id:req.body.userOfering});
+        const userOfering = await User.findById({_id:req.user.id});
         if (userOfering == null) res.status(404).json({error:"userOfering not found"});
       
         const userTaking = await User.findById({_id:req.body.userTaking});
         if (userTaking == null) res.status(404).json({error:"userTaking not found"});
 
-        if (req.body.userOfering == req.body.userTaking) res.status(404).json({error:"userTaking == userOfering"});
+        if (req.user.id == req.body.userTaking) res.status(404).json({error:"userTaking == userOfering"});
        
-        const productOfering = await Product.findById({_id:req.body.productOfering});
+        const productOfering = await Product.findById({_id:req.body.productOfering, userId: req.user.id});
         if (productOfering == null) res.status(404).json({error:"productOfering not found"});
 
-        const productTaking = await Product.findById({_id:req.body.productTaking});
+        const productTaking = await Product.findById({_id:req.body.productTaking, userId: req.body.userTaking});
         if (productTaking == null) res.status(404).json({error:"productTaking not found"}); 
-
-      //caldrà comprovar que es tenen prous punts i que el producte pertany a ususari dels dos ususaris
-     /* const userHasProduct = await User.hasProduct(_id:req.body.userOfering, product:req.body.product)
-      if (!userHasProduct) --> error 
-      userHasEnoughtPoints --> check ---> error
-      --> cal assignar points, que sigui obligatori i que al acabar el creat faci les sumes/restes corresponents als usuaris (sum_points i checkEnoguthPoints)
-      */
-  
+        
+        if (userTaking.points < tradeExchange.points) res.status(404).json({error:"not enought points"});
+        
         if (userOfering != null && userTaking != null && productOfering != null && productTaking  != null && req.body.userOfering != req.body.userTaking) {
-          const newProduct = await tradeExchange.save();
+          product.state = "reserved";
+          await product.save();
+          // UserTaking.getUserRewards();
+          // UserOfering.getUserRewards();
+          // UserTaking.getRewards();
+          // UserOfering.getRewards();
+          // UserTaking.levelManage();
+          // UserOfering.levelManage();
+          await tradeExchange.save();
           res.status(201).json(tradeExchange);  
      }
 
@@ -71,13 +74,13 @@ exports.readAllTradeExchange = async (req, res) => {
       console.log('Can not create the tradeExchange');
     }
 };
-  
-  exports.updateTradeExchange = async (req, res) => {
-  
-};
 
   exports.deleteTradeExchange = async (req, res) => {
-  
+    
+    if ('ADMIN' != req.user.role) {
+      res.status(401).json({error: "Do not have permission"})
+      return;
+    }
       try {
         const tradeExchange = await TradeExchange.findByIdAndDelete({_id: req.params.id});
